Move R8G8B8A8 struct over to a new Image3dStream header
Preparation before upcoming API change. There are no functional changes in this PR.
