+++ DummyLoader/Image3dSource.hpp
@@ -3,34 +3,8 @@ Designed by Fredrik Orderud <fredrik.orderud@ge.com>.
 Copyright (c) 2016, GE Healthcare, Ultrasound.      */
 #pragma once
 
-#define _USE_MATH_DEFINES // for M_PI
-#include <cmath>
-#include <memory>
-#include <array>
-#include <vector>
-#include "../Image3dAPI/ComSupport.hpp"
-#include "../Image3dAPI/IImage3d.h"
+#include "Image3dStream.hpp"
 
-#include "DummyLoader.h"
-#include "Resource.h"
-
-
-/** RGBA color struct that matches DXGI_FORMAT_R8G8B8A8_UNORM.
-Created due to the lack of such a class/struct in the Windows or Direct3D SDKs.
-Please remove this class if a more standardized alternative is available. */
-struct R8G8B8A8 {
-    R8G8B8A8 () : r(0), g(0), b(0), a(0) {
-    }
-
-    R8G8B8A8 (unsigned char _r, unsigned char _g, unsigned char _b, unsigned char _a) : r(_r), g(_g), b(_b), a(_a) {
-    }
-
-    operator unsigned int () const {
-        return *reinterpret_cast<const unsigned int*>(this);
-    }
-
-    unsigned char r, g, b, a;  ///< color channels
-};
 
 
 class ATL_NO_VTABLE Image3dSource :
